Calculate image size in floating-point instead of int.

+++ src/gifcreator.cpp
@@ -49,8 +49,8 @@ bool GIFCreator::createGIF(int frames_count, int frame_delay)
         QSize size = first_reader.size();
 
         if (!size.isEmpty()) {
-            if (size.width() * size.height() > GIF_MPIX_LIMIT * 1000000.0) {
-                qreal scale = qSqrt((size.width() * size.height()) / (GIF_MPIX_LIMIT * 1000000.0));
+            if (static_cast<qreal>(size.width()) * static_cast<qreal>(size.height()) > GIF_MPIX_LIMIT * 1000000.0) {
+                qreal scale = qSqrt((static_cast<qreal>(size.width()) * static_cast<qreal>(size.height())) / (GIF_MPIX_LIMIT * 1000000.0));
 
                 size.setWidth(qFloor(size.width()   / scale));
                 size.setHeight(qFloor(size.height() / scale));
